stacks/agents/app/pyproject.toml
[project]
name = "homelab-agent"
version = "0.1.0"
description = "Self-hosted AI agent service for PR review and coding tasks"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.28",
    "pydantic>=2",
    "pyjwt[crypto]>=2.9",
    "uvicorn[standard]>=0.34",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "ruff>=0.11",
    "ty>=0.0.1",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
